Added almost all CRUD actions for ProjectUrl model and created related views as well

+++ database/migrations/2020_03_17_095814_create_project_urls_table.php
@@ -16,7 +16,7 @@ class CreateProjectUrlsTable extends Migration
         Schema::create('project_urls', function (Blueprint $table) {
             $table->id();
             $table->string('url');
-            $table->integer('project_id')->unsigned();
+            $table->bigInteger('project_id')->unsigned();
             $table->timestamps();
         });
     }
